requirements.txt
flask==2.3.3
flask-socketio==5.3.6
# kafka-python==2.0.2
# confluent-kafka==2.2.0
# elasticsearch==8.8.0
textblob==0.17.1
vaderSentiment==3.3.2
plotly==5.15.0
pandas==2.0.3
python-dotenv==1.0.0

# pip install flask==2.3.3 flask-socketio==5.3.6 kafka-python==2.0.2 elasticsearch==8.8.0 textblob==0.17.1 vaderSentiment==3.3.2 plotly==5.15.0 pandas==2.0.3 python-dotenv==1.0.0


# python -m textblob.download_corpora
